CLR proxy handles are now long

+++ trunk/jni4net.j/src/main/java/java_/io/Serializable_.java
@@ -29,7 +29,7 @@ public final class Serializable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Serializable extends system.Object implements java.io.Serializable {
     
-    protected __Serializable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Serializable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
 }
