+++ trunk/jni4net.j/src/main/java/java_/lang/Cloneable_.java
@@ -29,7 +29,7 @@ public final class Cloneable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Cloneable extends system.Object implements java.lang.Cloneable {
     
-    protected __Cloneable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Cloneable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
 }
