+++ trunk/jni4net.j/src/main/java/java_/lang/Runnable_.java
@@ -29,7 +29,7 @@ public final class Runnable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __Runnable extends system.Object implements java.lang.Runnable {
     
-    protected __Runnable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __Runnable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
