+++ trunk/jni4net.j/src/main/java/system/IDisposable_.java
@@ -29,7 +29,7 @@ public final class IDisposable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __IDisposable extends system.Object implements system.IDisposable {
     
-    protected __IDisposable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __IDisposable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
