+++ trunk/jni4net.j/src/main/java/system/IFormattable_.java
@@ -29,7 +29,7 @@ public final class IFormattable_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __IFormattable extends system.Object implements system.IFormattable {
     
-    protected __IFormattable(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __IFormattable(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
