+++ trunk/jni4net.j/src/main/java/system/InvalidOperationException.java
@@ -15,7 +15,7 @@ public class InvalidOperationException extends system.SystemException {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected InvalidOperationException(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected InvalidOperationException(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
