+++ trunk/jni4net.j/src/main/java/system/collections/ICollection_.java
@@ -29,7 +29,7 @@ public final class ICollection_ {
 @net.sf.jni4net.attributes.ClrProxy
 class __ICollection extends system.Object implements system.collections.ICollection {
     
-    protected __ICollection(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected __ICollection(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
