+++ trunk/jni4net.j/src/main/java/system/io/Stream.java
@@ -15,7 +15,7 @@ public class Stream extends system.MarshalByRefObject implements system.IDisposa
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected Stream(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected Stream(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
