+++ trunk/jni4net.j/src/main/java/system/reflection/BindingFlags.java
@@ -15,7 +15,7 @@ public class BindingFlags extends system.Enum {
     //<generated-proxy>
     private static system.Type staticType;
     
-    protected BindingFlags(net.sf.jni4net.inj.INJEnv __env, int __handle) {
+    protected BindingFlags(net.sf.jni4net.inj.INJEnv __env, long __handle) {
             super(__env, __handle);
     }
